pyproject.toml
[project]
name = "gmdm"
version = "0.2.5"
authors = [{name = "Kenan Masri", email = "kenanmasri@outlook.com"}]
description = "The GMDM CLI tool for managing dependencies of GameMaker projects and files."
readme = "README.md"
requires-python = ">=3.11"
keywords = ["gamemaker", "gml", "dependencies", "export", "libraries",]
license = {text = "MIT License"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "coloredlogs",
    "PyYAML",
    "json5",
]
dynamic = []

[project.scripts]
gmdm = "gmdm.cli:main"

[project.urls]
"Bug Tracker" = "https://github.com/knno/gmdm/issues"